Loading table data from xml files into arbitrary classes dynamically
Recipe (recept) blazor component now runs general code with just two string parameters, which fetches xml files in a recipe subdirectory, then returns as object list which the recipe page casts as the recipe class and loads into data table.

Additional feature: for now, the first 4 properties in the class being used with the DataTable component automatically get used as columns, making it easy to change. Later it would be good with a UI option to filter columns.